NETOBSERV-2092: integrate network events with drops (#922)
* NETOBSERV-2092: integrate network events with drops

- network events metrics now include bytes/packets as u16
  - decrease bytes/packets to u16 in drops, to limit the memory increase; assume that metric is capped
    to ~65000
  - manage overflow (maxed at ~65000)
- After decoding network events, if it's a drop, it's added to the drop
  metric (bytes, packet, cause)
  - New drop causes for network events, based on ovnk-observ "actor"
    field
- Add tests

* Address feedback

* Clarify structured/unstructure netevents map decode

* regen bytecode

+++ pkg/pbflow/proto.go
@@ -87,7 +87,7 @@ func FlowToPB(fr *model.Record) *Record {
 		}
 	}
 	if fr.Metrics.PktDropMetrics != nil {
-		pbflowRecord.PktDropBytes = fr.Metrics.PktDropMetrics.Bytes
+		pbflowRecord.PktDropBytes = uint64(fr.Metrics.PktDropMetrics.Bytes)
 		pbflowRecord.PktDropPackets = uint64(fr.Metrics.PktDropMetrics.Packets)
 		pbflowRecord.PktDropLatestFlags = uint32(fr.Metrics.PktDropMetrics.LatestFlags)
 		pbflowRecord.PktDropLatestState = uint32(fr.Metrics.PktDropMetrics.LatestState)
@@ -178,8 +178,8 @@ func PBToFlow(pb *Record) *model.Record {
 				Latency: uint64(pb.DnsLatency.AsDuration()),
 			},
 			PktDropMetrics: &ebpf.BpfPktDropMetrics{
-				Bytes:           pb.PktDropBytes,
-				Packets:         uint32(pb.PktDropPackets),
+				Bytes:           uint16(pb.PktDropBytes),
+				Packets:         uint16(pb.PktDropPackets),
 				LatestFlags:     uint16(pb.PktDropLatestFlags),
 				LatestState:     uint8(pb.PktDropLatestState),
 				LatestDropCause: pb.PktDropLatestDropCause,
